+++ R/archivist-package.R
@@ -56,6 +56,7 @@
 #' @importMethodsFrom git2r add
 #' @importMethodsFrom git2r commit
 #' @importMethodsFrom git2r push
+#' @importMethodsFrom git2r pull
 #' @importMethodsFrom git2r init
 #' @importMethodsFrom git2r remote_add
 #' @importMethodsFrom git2r clone